Revert "Create check availability (#3)"
This reverts commit 30a78cbb1e8e95987b307d057ccc004fb4c73a16.
