+++ ormconfig.example.env
@@ -3,5 +3,6 @@ TYPEORM_HOST = localhost
 TYPEORM_USERNAME = TYPEORM_USERNAME
 TYPEORM_PASSWORD = TYPEORM_PASSWORD
 TYPEORM_DATABASE = TYPEORM_DATABASE
+TYPEORM_SCHEMA = TYPEORM_SCHEMA
 TYPEORM_PORT = 5432
-TYPEORM_SYNCHRONIZE = false
+TYPEORM_SYNCHRONIZE = true
