+++ src/app.module.ts
@@ -1,10 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { getConnectionOptions } from 'typeorm';
-import { AvailabilityModule } from './modules/availability.module';
-import { BuildingModule } from './modules/building.module';
-import { PropertyModule } from './modules/property.module';
-import { ReservationModule } from './modules/reservation.module';
 
 @Module({
   imports: [
@@ -14,10 +10,6 @@ import { ReservationModule } from './modules/reservation.module';
           autoLoadEntities: true,
         }),
     }),
-    AvailabilityModule,
-    BuildingModule,
-    PropertyModule,
-    ReservationModule,
   ],
 })
 export class AppModule {}
