--- src/dtos/date-filter.dto.ts
@@ -1,11 +1,0 @@
-import { IsDateString, IsNotEmpty } from 'class-validator';
-
-export class DateFilterDto {
-  @IsNotEmpty()
-  @IsDateString({ strict: true })
-  start: string;
-
-  @IsNotEmpty()
-  @IsDateString({ strict: true })
-  end: string;
-}
