--- src/dtos/flexible-filter.dto.ts
@@ -1,21 +1,0 @@
-import {
-  ArrayNotEmpty,
-  IsArray,
-  IsEnum,
-  IsIn,
-  IsNotEmpty,
-} from 'class-validator';
-import { FlexibilityType } from 'src/enums/flexibility-type.enum';
-import { months } from 'src/utils/constants';
-
-export class FlexibleFilterDto {
-  @IsNotEmpty()
-  @IsEnum(FlexibilityType)
-  type: FlexibilityType;
-
-  @IsNotEmpty()
-  @IsArray()
-  @ArrayNotEmpty()
-  @IsIn(Object.keys(months), { each: true })
-  months: string[];
-}
