--- src/enums/amenities.enum.ts
@@ -1,7 +1,0 @@
-export enum Amenities {
-  WiFi = 'WiFi',
-  Pool = 'Pool',
-  Garden = 'Garden',
-  Parking = 'Parking',
-  'Tennis table' = 'Tennis table',
-}
