--- src/enums/flexibility-type.enum.ts
@@ -1,5 +1,0 @@
-export enum FlexibilityType {
-  weekend = 'weekend',
-  week = 'week',
-  month = 'month',
-}
