--- src/enums/property-type.enum.ts
@@ -1,5 +1,0 @@
-export enum PropertyType {
-  '1bdr' = '1bdr',
-  '2bdr' = '2bdr',
-  '3bdr' = '3bdr',
-}
