--- src/exceptions/MultipleTimeFilters.exception.ts
@@ -1,10 +1,0 @@
-import { HttpException, HttpStatus } from '@nestjs/common';
-
-export class MultipleTimeFiltersException extends HttpException {
-  constructor() {
-    super(
-      'Please provide either date or flexibility',
-      HttpStatus.NOT_ACCEPTABLE,
-    );
-  }
-}
