--- src/exceptions/NoTimeFilters.exception.ts
@@ -1,10 +1,0 @@
-import { HttpException, HttpStatus } from '@nestjs/common';
-
-export class NoTimeFiltersException extends HttpException {
-  constructor() {
-    super(
-      'Please provide at least date or flexibility',
-      HttpStatus.NOT_ACCEPTABLE,
-    );
-  }
-}
