--- src/interfaces/property-availability-check-result.interface.ts
@@ -1,10 +1,0 @@
-export interface IAlternative {
-  id: number;
-  availableStarting: string;
-}
-
-export interface IPropertyAvailabilityCheckResult {
-  match: number[] | [];
-  alternative: IAlternative[] | [];
-  other: number[] | [];
-}
