--- src/modules/availability.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { AvailabilityRepository } from '../repositories/availability.repository';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([AvailabilityRepository])],
-})
-export class AvailabilityModule {}
