--- src/modules/reservation.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { ReservationRepository } from 'src/repositories/reservation.repository';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([ReservationRepository])],
-})
-export class ReservationModule {}
