+++ src/property/property.model.ts
@@ -1,0 +1,13 @@
+export enum PropertyType {
+  '1bdr' = '1bdr',
+  '2bdr' = '2bdr',
+  '3bdr' = '3bdr',
+}
+
+export enum PropertyAmenities {
+  WiFi = 'WiFi',
+  Pool = 'Pool',
+  Garden = 'Garden',
+  Parking = 'Parking',
+  'Tennis table' = 'Tennis table',
+}
