--- src/repositories/availability.repository.ts
@@ -1,5 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { Availability } from '../entities/availability.entity';
-
-@EntityRepository(Availability)
-export class AvailabilityRepository extends Repository<Availability> {}
