--- src/repositories/building.repository.ts
@@ -1,5 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { Building } from '../entities/building.entity';
-
-@EntityRepository(Building)
-export class BuildingRepository extends Repository<Building> {}
