--- src/repositories/property.repository.ts
@@ -1,5 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { Property } from '../entities/property.entity';
-
-@EntityRepository(Property)
-export class PropertyRepository extends Repository<Property> {}
