--- src/repositories/reservation.repository.ts
@@ -1,5 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { Reservation } from '../entities/reservation.entity';
-
-@EntityRepository(Reservation)
-export class ReservationRepository extends Repository<Reservation> {}
