--- src/utils/constants.ts
@@ -1,22 +1,0 @@
-export const months: { [key: string]: number } = {
-  jan: 1,
-  feb: 2,
-  mar: 3,
-  apr: 4,
-  may: 5,
-  jun: 6,
-  jul: 7,
-  aug: 8,
-  sep: 9,
-  oct: 10,
-  nov: 11,
-  dec: 12,
-};
-
-export const fridaySaturdayWeekendCities = ['Dubai'];
-
-export const flexibilityIncrementDays = {
-  weekend: 1,
-  week: 6,
-  month: 29,
-};
